+++ app/Http/Requests/FarmStore.php
@@ -6,9 +6,11 @@ use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Foundation\Http\FormRequest;
 use App\Exceptions\MissingInputException;
 use Illuminate\Validation\Rule;
+use App\Http\Requests\Concerns\FarmValidationErrorsHandler;
 
 class FarmStore extends FormRequest
 {
+    use FarmValidationErrorsHandler;
     /**
      * Determine if the user is authorized to make this request.
      *
@@ -34,28 +36,6 @@ class FarmStore extends FormRequest
             'capacity' => 'required|numeric|integer',
             'launched_date' => 'required|date|date_format:Y-m-d',
             'status' => 'required'
-        ];
-    }
-
-    protected function failedValidation(Validator $validator)
-    {
-        $missingInputs = [];
-        foreach ($this->rules() as $input => $rules) {
-            if ($validator->errors()->has($input)) {
-                $missingInputs[$input] = $validator->errors()->first($input);
-            }
-        }
-        throw new MissingInputException($missingInputs);
-    }
-
-    public function messages()
-    {
-        return [
-            'name.required' => 'Farm name field is required',
-            'name.regex' => 'Farm name field must only contain alpha-numeric characters',
-            'name.string' => 'Farm name field must be string',
-            'address.required' => 'Farm address field is required',
-            'coordinates.string' => 'The coordinates field must be a string.',
         ];
     }
 }
